refactor: restructure project directories
