+++ Back-End/src/controllers/userController.js
@@ -1,4 +1,4 @@
-const manager = require('../managers/userManager')
+const manager = require('../models/userManager')
 
 const getUser = async login => {
     return manager.findByLogin(login)
